Initial commit of assg05 solution
Initial commit of a full solution for assignment 05.  Update
assignment description, and unit tests.  Implementation of the
SchedulingSystem functions (tasks 1-4) are in this commit.

All created unit tests are passing for this example solution.
Remaining is to implement scheduling policies, update
system tests, and check second half of assignment instructions.